added exceptions and working get, get rid of starting page and refactor, documenation FE: added button

+++ SocialAnalyser/Middlewares/ExceptionMiddleware.cs
@@ -32,7 +32,7 @@ namespace SocialAnalyser.Middlewares
     private Task HandleExceptionAsync(HttpContext context, Exception exception)
     {
       ApiError apiError = fApiErrorFactory.Create(exception);
-      context.Response.StatusCode = (int)apiError.StatusCode;
+      context.Response.StatusCode = apiError.StatusCode;
       context.Response.ContentType = "application/json";
 
       return context.Response.WriteAsync(JsonConvert.SerializeObject(new { message = apiError.Message }));
